secured routes

+++ routes/product.js
@@ -2,6 +2,7 @@ const express = require('express');
 const bodyParser = require('body-parser');
 const controllers = require('../controllers');
 const ProductController = controllers.ProductController;
+const UserController = controllers.UserController;
 
 const productRouter = express.Router();
 productRouter.use(bodyParser.json());
@@ -15,14 +16,19 @@ productRouter.post('/', function(req, res) {
         res.status(400).end();
         return;
     }
-    ProductController.add(name, style, description, parseFloat(price))
-        .then((product) => {
-            res.status(201).json(product);
-        })
-        .catch((err) => {
-            console.error(err);
-            res.status(500).end();
-        });
+    if(UserController.verifyToken(req.headers['x-access-token']) === true){
+        ProductController.add(name, style, description, parseFloat(price))
+            .then((product) => {
+                res.status(201).json(product);
+            })
+            .catch((err) => {
+                console.error(err);
+                res.status(500).end();
+            });
+    }else{
+        res.status(403).end();
+    }
+
 });
 
 productRouter.post('/update', function(req, res) {
@@ -32,14 +38,19 @@ productRouter.post('/update', function(req, res) {
         res.status(400).end();
         return;
     }
-    ProductController.update(name, parseFloat(price))
-        .then((product) => {
-            res.status(201).json({"New price" : price});
-        })
-        .catch((err) => {
-            console.error(err);
-            res.status(500).end();
-        });
+    if(UserController.verifyToken(req.headers['x-access-token']) === true){
+        ProductController.update(name, parseFloat(price))
+            .then((product) => {
+                res.status(201).json({"New price" : price});
+            })
+            .catch((err) => {
+                console.error(err);
+                res.status(500).end();
+            });
+    }else{
+        res.status(403).end();
+    }
+
 });
 
 productRouter.get('/', function (req, res) {
@@ -57,13 +68,18 @@ productRouter.get('/', function (req, res) {
 
 productRouter.delete('/:id', function (req, res) {
     const productId = parseInt(req.params.id);
-    ProductController.delete(productId)
-        .then((product) => {
-            res.status(200).json(product.deletedAt);
-        })
-        .catch((err) => {
-            res.status(500).end();
-        });
+    if(UserController.verifyToken(req.headers['x-access-token']) === true){
+        ProductController.delete(productId)
+            .then((product) => {
+                res.status(200).json(product.deletedAt);
+            })
+            .catch((err) => {
+                res.status(500).end();
+            });
+    }else{
+        res.status(403).end();
+    }
+
 });
 
 module.exports = productRouter;
